refactor and fix ESlint errors
